+++ common/model/src/main/java/is/codion/common/model/selection/SingleSelectionModel.java
@@ -44,4 +44,9 @@ public interface SingleSelectionModel<R> {
 	 * @return a {@link Mutable} controlling the selected item
 	 */
 	Mutable<R> item();
+
+	/**
+	 * Clears the selection
+	 */
+	void clear();
 }